Add files via upload
In this project, we have tried to implement the basics of computer graphics using the mid-point line drawing algorithm and circle drawing algorithm. We also used transformation, rotation, and scaling here. The concept of our project is a half-filled glass is there. We used rotation to rotate the glass and scaling to increase the water level. We used mid-point line drawing algorithm for drawing the class and mid-point circle drawing algorithm to draw the water drops. Taking user inputs all the actions will take place.